Add checks for readable and writable atoms
Previously we had (atom-like?) fn for readable atoms,
and (satisfies? IAtom x) for writable atoms. This commit
adds two functions, (readable-atom-like? x) and
(writable-atom-like? x). The main motivation for adding these
checks is that in some situations we may want to provide initial
data in an atom-like thing that is read-only. Devcards
shouldn't really need the atom to be writable unless :history
is set to true, so we should be able to accept read-only atoms.